undo dir and not clean log removal
- as requested in code review, will go with a safety-first operating assumption

- as such, with safety as modus operandus, make the logging more detailed in these scenarios, since they're not supposed to happen
  - and don't check code coverage on these as they're not supposed to happen in normal usage

+++ src/tscache.ts
@@ -153,9 +153,24 @@ export class TsCache
 			return;
 
 		const entries = fs.readdirSync(this.cacheRoot);
-		entries.filter(e => e.startsWith(this.cachePrefix)).forEach((e) =>
+		entries.forEach((e) =>
 		{
 			const dir = `${this.cacheRoot}/${e}`;
+
+			/* istanbul ignore if -- this is a safety check, but shouldn't happen when using a dedicated cache dir */
+			if (!e.startsWith(this.cachePrefix))
+			{
+				this.context.debug(`skipping cleaning ${dir} as it does not have prefix ${this.cachePrefix}`);
+				return;
+			}
+
+			/* istanbul ignore if -- this is a safety check, but should never happen in normal usage */
+			if (!fs.statSync(dir).isDirectory)
+			{
+				this.context.debug(`skipping cleaning ${dir} as it is not a directory`);
+				return;
+			}
+
 			this.context.info(blue(`cleaning cache: ${dir}`));
 			fs.removeSync(`${dir}`);
 		});
